requirements.txt
firebase-admin>=6.0.0
joblib>=1.2.0
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
python-dotenv>=0.20.0
scikit-learn>=1.1.0
